Add USB CDC Recieve ring buffer, add BDEAD command to jump system bootloader.

+++ Middlewares/ST/STM32_USB_Device_Library/Core/Inc/usbd_core.h
@@ -140,8 +140,8 @@ USBD_StatusTypeDef  USBD_LL_PrepareReceive(USBD_HandleTypeDef *pdev,
                                            uint8_t  *pbuf,
                                            uint16_t  size);
 
-uint32_t USBD_LL_GetRxDataSize  (USBD_HandleTypeDef *pdev, uint8_t  ep_addr);  
-void  USBD_LL_Delay (uint32_t Delay);
+uint16_t USBD_LL_GetRxDataSize  (USBD_HandleTypeDef *pdev, uint8_t  ep_addr);  
+void USBD_LL_Delay (uint32_t Delay);
 
 /**
   * @}
